backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
pdfminer.six>=20231228
python-docx>=1.1.0
